+++ apollo-client/src/main/java/com/ctrip/framework/apollo/spi/DefaultConfigFactory.java
@@ -43,7 +43,7 @@ import com.ctrip.framework.apollo.util.ConfigUtil;
  */
 public class DefaultConfigFactory implements ConfigFactory {
   private static final Logger logger = LoggerFactory.getLogger(DefaultConfigFactory.class);
-  private ConfigUtil m_configUtil;
+  private final ConfigUtil m_configUtil;
 
   public DefaultConfigFactory() {
     m_configUtil = ApolloInjector.getInstance(ConfigUtil.class);
@@ -53,9 +53,13 @@ public class DefaultConfigFactory implements ConfigFactory {
   public Config create(String namespace) {
     ConfigFileFormat format = determineFileFormat(namespace);
     if (ConfigFileFormat.isPropertiesCompatible(format)) {
-      return new DefaultConfig(namespace, createPropertiesCompatibleFileConfigRepository(namespace, format));
+      return this.createRepositoryConfig(namespace, createPropertiesCompatibleFileConfigRepository(namespace, format));
     }
-    return new DefaultConfig(namespace, createLocalConfigRepository(namespace));
+    return this.createRepositoryConfig(namespace, createLocalConfigRepository(namespace));
+  }
+
+  protected Config createRepositoryConfig(String namespace, ConfigRepository configRepository) {
+    return new DefaultConfig(namespace, configRepository);
   }
 
   @Override